refactor(consolidate): extract plan-merge algebra to consolidate/merge (D3 step 5)
Move isValidOp + mergePlans into consolidate/merge.ts — pure op-reconciliation
algebra over ConsolidateOperation arrays (no fs/LLM/config). Drops the now
body-unused op sub-type imports from the orchestrator (kept in the re-export).
Re-exported mergePlans; behavior-identical. 252 consolidate tests green.
